+++ CHANGELOG.md
@@ -18,6 +18,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 * #1273 Add caching of loading of OIDC private key.
+* #1285 Add post_logout_redirect_uris field in application views.
 
 - ### Fixed
 * #1284 Allow to logout whith no id_token_hint even if the browser session already expired